Safer data handling around phone number
This script previously assumed that a user would have at least one of
their first two address entries with a phone number. Because of this,
the script would crash if it came upon a user without a phone number (or
at least, without a phone number in the first two entries).

This change iterates through the entire list of addresses and finds the
first entry with a phone number, then uses that. If no phone numbers are
found, it just uses the string "N/A".