Major refactor to make testing easier
This refactors a lot of the code in the existing dailyClassReminder.py
script to make it more readable and easier to add tests. Mostly, I've
deleted unused code and simplified data structures where possible (e.g.
where some parts of the data are never referenced).

This is a pretty major refactor, so I wouldn't expect to port this in
all in one PR. I wanted to put something together to show how a refactor
could make the code more readable and easier to work with - that said,
there are a ton of edge cases and weird behaviors that I'm not familiar
with, so I want to make sure that we take it step-by-step to make
changes and reduce the risk of these scripts breaking.