Add test infrastructure for openPathUpdateAll
This adds two main pieces of code - the MockNeonUserBuilder and the
MockAltaUserBuilder. The Builder pattern allows us to create complex
objects (like fake user accounts) in a relatively readable way, and we
can then use those mock accounts to test out the behavior of real
functions without needing to connect to either Alta or Neon.

This also adds a test for the openPathUpdateAll method, but I'll be
working on cleaning that up to make it more readable.